Dynamic dispatch for cursor swap and draw

+++ src/rect.h
@@ -5,10 +5,10 @@
 #include "stdint.h"
 
 struct IRect{
-	uint8_t x;
-	uint8_t y;
-	uint8_t w;
-	uint8_t h;
+	int16_t x;
+	int16_t y;
+	int16_t w;
+	int16_t h;
 };
 
 #endif // rect_h_INCLUDED
